requirements.txt
absl-py==0.1.11
astor==0.6.2
bleach==1.5.0
boto==2.48.0
boto3==1.5.16
botocore==1.8.30
bz2file==0.98
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
docutils==0.14
et-xmlfile==1.0.1
gast==0.2.0
gensim==3.2.0
grpcio==1.10.0
html5lib==0.9999999
idna==2.6
jdcal==1.3
jieba==0.39
jmespath==0.9.3
Keras==2.1.5
Markdown==2.6.11
matplotlib==2.1.1
nltk==3.2.5
numpy==1.14.1
openpyxl==2.4.9
pandas==0.22.0
Pillow==5.0.0
protobuf==3.5.2
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
smart-open==1.5.6
snownlp==0.12.3
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
xlrd==1.1.0
